+++ flutter_application_1/lib/Interfaz/InicioSesion/Estilo/index.dart
@@ -1,4 +1,4 @@
-export 'boton.dart';
+export 'botones.dart';
 export 'containerError.dart';
 export 'containerLabelForm.dart';
 export 'containerTitle.dart';
