+++ flutter_application_1/lib/Interfaz/InicioSesion/index.dart
@@ -4,4 +4,4 @@ export 'inicioApp.dart';
 export 'registrarse1.dart';
 export 'registrarse2.dart';
 export 'registrarse3.dart';
-export 'registrarse4.dart';
+export 'confirmarRegistro.dart';
